Added filtering of snippets with assessment in step 5
Run python3 filter.templated.py -h to get help on input arguments.
Snippets that are deemed to contain terms that indicate assessments
are dropped from the allPos and allNeg files. Note that the entire
snippet is dropped, not just the assesment terms. For now, the
output of the filter is set to produce allPos_filtered.txt and
allNeg_filtered.txt.
Two files have been added to the resources: one that contains terms
that suggest presence of assesment language and one that contains
other terms that should lead to dropping a snippet. Specifying the
latter file is optional. It can be used, for example, to delete
snippets that are known to often be mislabelled.