New file names for step 4 output.
File names for step 4 output and step 5 input and output were
changed conform agreement to keep the CLEVER output file names
unchanged for the version that includes step 5. This means that
the final files are still called allPos.txt and allNeg.txt and
no downstream changes are necessary for a pipeline that uses
CLEVER output. The output of step 4 now includes 'unfiltered' in
the file names.
A bash script was added for step 5 that includes all targets that
are listed for the other steps.