Offset field extended with third field for main target.
The note and snippet offset field was extended to indicate the offset
in the snippet for the main target term that triggered the snippet.
This information is used by other apps (e.g. MAVA). In order to make
this work, context terms now include all terms from the dictionary,
instead of all terms except thos belonging to the main target.
Removal of the first and last token from the snippet was moved back
to step 4 (from step 5) and is now conditional on those tokens not
being part of the term that was tagged.
Contents of the variables noffset and termid were switched. This was
corrected. Test notes have been updated to include tests for the
correct offset of the main target term in the snippet if that term
happens to be mentioned more than once in the snippet.